backend/requirements.txt
# FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Database dependencies
sqlalchemy==1.4.53
psycopg2-binary==2.9.7
alembic==1.12.1
databases[postgresql]==0.8.0

# Validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# Environment and configuration
python-dotenv==1.0.0

# Security and CORS - handled by FastAPI middleware

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
requests==2.31.0 